+++ app/src/main/java/com/forrestguice/suntimeswidget/settings/WidgetSettings.java
@@ -33,6 +33,10 @@ import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_1;
 import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_2;
 import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_3;
 import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_4;
+import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_5;
+import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_6;
+import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_7;
+import com.forrestguice.suntimeswidget.layouts.MoonLayout_1x1_8;
 import com.forrestguice.suntimeswidget.layouts.SunLayout;
 import com.forrestguice.suntimeswidget.layouts.SunLayout_1x1_0;
 import com.forrestguice.suntimeswidget.layouts.SunLayout_1x1_1;
@@ -402,7 +406,11 @@ public class WidgetSettings
         MODE1x1_PHASEILLUM("Moon phase & illumination", R.layout.layout_widget_moon_1x1_1),
         MODE1x1_PHASE("Moon phase only", R.layout.layout_widget_moon_1x1_2),
         MODE1x1_ILLUM("Moon illumination only", R.layout.layout_widget_moon_1x1_3),
-        MODE1x1_PHASENEXT("Next major phase", R.layout.layout_widget_moon_1x1_4);
+        MODE1x1_PHASENEXT("Next major phase", R.layout.layout_widget_moon_1x1_4),
+        MODE1x1_ALTAZ("Altitude & Azimuth", R.layout.layout_widget_moon_1x1_5),
+        MODE1x1_DECRIGHT("Declination & Right Ascension", R.layout.layout_widget_moon_1x1_6),
+        MODE1x1_DISTANCE("Current distance", R.layout.layout_widget_moon_1x1_7),
+        MODE1x1_APSIS("Next apogee / perigee", R.layout.layout_widget_moon_1x1_8);
 
         private final int layoutID;
         private String displayString;
@@ -440,6 +448,10 @@ public class WidgetSettings
             MODE1x1_PHASE.setDisplayString(context.getString(R.string.widgetMode1x1_moonphase));
             MODE1x1_ILLUM.setDisplayString(context.getString(R.string.widgetMode1x1_moonillum));
             MODE1x1_PHASENEXT.setDisplayString(context.getString(R.string.widgetMode1x1_moonphasenext));
+            MODE1x1_ALTAZ.setDisplayString(context.getString(R.string.widgetMode1x1_altaz));
+            MODE1x1_DECRIGHT.setDisplayString(context.getString(R.string.widgetMode1x1_decright));
+            MODE1x1_DISTANCE.setDisplayString(context.getString(R.string.widgetMode1x1_distance));
+            MODE1x1_APSIS.setDisplayString(context.getString(R.string.widgetMode1x1_apsis));
         }
     }
 
@@ -1214,6 +1226,22 @@ public class WidgetSettings
         WidgetModeMoon1x1 mode = loadMoon1x1ModePref(context, appWidgetId);
         switch (mode)
         {
+            case MODE1x1_APSIS:
+                layout = new MoonLayout_1x1_8();
+                break;
+
+            case MODE1x1_DISTANCE:
+                layout = new MoonLayout_1x1_7();
+                break;
+
+            case MODE1x1_DECRIGHT:
+                layout = new MoonLayout_1x1_6();
+                break;
+
+            case MODE1x1_ALTAZ:
+                layout = new MoonLayout_1x1_5();
+                break;
+
             case MODE1x1_PHASENEXT:
                 layout = new MoonLayout_1x1_4();
                 break;